create module for  get_tournament_round
